Layer 135

+++ src/CatraProto.Client/TL/Schemas/CloudChats/BotInfoBase.cs
@@ -8,7 +8,7 @@ namespace CatraProto.Client.TL.Schemas.CloudChats
     {
 
 [Newtonsoft.Json.JsonProperty("user_id")]
-		public abstract int UserId { get; set; }
+		public abstract long UserId { get; set; }
 
 [Newtonsoft.Json.JsonProperty("description")]
 		public abstract string Description { get; set; }
